moves ethereum-integrated pre-boot validation to vere/dawn
